+++ src/main/java/com/mihailTs/trading_bot/model/TrainingAsset.java
@@ -4,12 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @EqualsAndHashCode(callSuper = true)
 @Data
 public class TrainingAsset extends Asset{
-    public TrainingAsset(String tokenId, double quantity, LocalDateTime createdAt, LocalDateTime updatedAt) {
+    public TrainingAsset(String tokenId, BigDecimal quantity, LocalDateTime createdAt, LocalDateTime updatedAt) {
         super(tokenId, quantity, createdAt, updatedAt);
     }
 }
